compile warnings

+++ src/openms/include/OpenMS/KERNEL/MSExperiment.h
@@ -546,10 +546,10 @@ public:
     void clear(bool clear_meta_data);
 
     /// returns true if at least one of the spectra has the specified level
-    bool containsScanOfLevel(int ms_level) const;
+    bool containsScanOfLevel(size_t ms_level) const;
 
     /// returns true if any MS spectra of the specified level contain at least one peak with intensity of 0.0
-    bool hasZeroIntensities(int ms_level) const;
+    bool hasZeroIntensities(size_t ms_level) const;
 
 
 protected:
